% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_load.R
\name{initial_load}
\alias{initial_load}
\title{Initial loading of spectra}
\usage{
initial_load(raw.spec = NULL, file = NULL, corr.norm, use.eshift = NULL)
}
\arguments{
\item{raw.spec}{List of files already loaded with read_raw_spec()}

\item{file}{Vector with file names}

\item{corr.norm}{Initial correction and normalization parameters}

\item{use.eshift}{Set TRUE, if using energy shift value, defaults to NULL}
}
\description{
This function loads and initially normalizes the raw spectra. Output is a list with the raw and initially corrected spectra.
}
\examples{
data(stdmix)
corr.spec.standards  <- initial_load(specdat[1:4], 
  corr.norm = c(-36, -15, 37, 58))
corr.spec.samples    <- initial_load(specdat[5:8], 
  corr.norm = c(-36, -15, 37, 58))
}
\keyword{correction}
\keyword{normalization,}
