% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descr.R
\docType{data}
\name{specdat}
\alias{specdat}
\title{Phosphorus K-edge XANES spectral data for LCF}
\format{List of four standard spectra comounds and four spectra of mixtures of these standard compounds, output objects of function read_raw_spec().}
\usage{
data(stdmix)
}
\description{
Data from an experiment on the quality of XANES spectrscopy for phosphorus speciation
}
\examples{
data(stdmix)
specdat[[1]]
specdat[[5]]
}
\references{
Werner & Prietzel 2015, Environ. Sci. Technol. (49), 10521-10528
(\href{dx.doi.org/10.1021/acs.est.5b03096}{DOI})
}
\keyword{datasets}
