% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_spec.R
\name{read_raw_spec}
\alias{read_raw_spec}
\title{Read raw spectra}
\usage{
read_raw_spec(file)
}
\arguments{
\item{file}{The raw .xmu file}
}
\description{
This function reads the raw .xmu file, extracts E0 and returns a list with name, E0 and the raw spectrum
}
\examples{
## any .xmu file as output from ATHENA (>=0.9.25)
}
\keyword{correction}
\keyword{normalization,}

